Release 1.4.7-part2: modular memory store, fixes & security
